feat: Add Memory Game randomized deck creation
Hard coded card li elements were removed from the index page to allow
for randomized card addition using JavaScript.

A self executing function is used to generate a randomized deck of
cards that are added to the deck class page element. The self
executing function takes a list of unique game cards, creates a pair
card for each card argument and generates an il element for each.

The li elements are held in a document fragment and added to The
DOM once the entire deck is created, resulting in a single browser
redraw cycle for the entire deck.