feat: Add game completion modal
The game end modal is placed on top of a fullscreen overlay to create
visual focus. Player game statistics are displayed and the player is
presented with a PLAY AGAIN button.

The modal is not displayed until the endGame function is invoked upon
completion of a round by the user.