feat: Add Player class score and position reset
Game score tracking and counting takes place within the update method,
which is continually runnning within the game Engine. If the player
reaches the water a game score is counted.

The player position restart is handled by the Player's reset method
which is run whenever the palyer reaches the water. This method will
also be used in restarting the player on collision with the Enemey
objects in game in future features.