feat: Add site assets to service worker cache
Added the remaining site assets to the service worker in order to cache
ALL content. Restaurant details pages that have been visited at least
once will fully render if the user is offline. If the restaurant detail
page was not fully cached, it will render while offline but will not
have a mapbox interactive map.